requirements.txt
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
asgiref
attrs
certifi
charset-normalizer
dj-database-url
Django
djangorestframework
djangorestframework-simplejwt
frozenlist
gunicorn
idna
multidict
packaging
propcache
PyJWT
python-decouple
requests
sqlparse
twilio
typing_extensions
tzdata
urllib3
yarl
psycopg2
psycopg2-binary>=2.9.9
cloudinary
djangorestframework-gis
channels
redis
daphne


